+++ Castlevania/GCamera.h
@@ -13,10 +13,10 @@ public:
 	RECT GetRegion(float left,float top, float width,float height);
 	GCamera();
 	D3DXVECTOR3 CenterSprite(int x, int y);	//Tìm tọa độ center của sprite
-	void UpdateCamera(int x);		
+	void UpdateCamera(float x);		
 	void UpdateCamera(int &w, int &h);
 	void SetSizeMap(int _max, int _min);
-	D3DXVECTOR2 Transform(int x, int y);		//Hàm transform, chuyển tọa độ viewport về tọa độ world
+	D3DXVECTOR2 Transform(float x, float y);		//Hàm transform, chuyển tọa độ viewport về tọa độ world
 	void setViewPort(D3DXVECTOR2 view);
 };
